fix: remove `impl From<usize> for MemoryValue`

+++ acvm-repo/brillig_vm/tests/foreign_calls.rs
@@ -526,12 +526,12 @@ fn foreign_call_opcode_nested_arrays_input() {
     let mut memory = vec![MemoryValue::from(1_u32)];
 
     // Declare a4: [RC, ...items]
-    let a4_ptr = memory.len();
+    let a4_ptr = memory.len() as u32;
     memory.extend(vec![MemoryValue::from(1_u32)]);
     memory.extend(a4.clone());
 
     // Declare a9: [RC, ...items]
-    let a9_ptr = memory.len();
+    let a9_ptr = memory.len() as u32;
     memory.extend(vec![MemoryValue::from(1_u32)]);
     memory.extend(a9.clone());
